hai update statistic
