+++ meeting-room-management/src/app/statistic-room/model/booked-chart.class.ts
