+++ meeting-room-management/src/app/statistic-room/model/booked-room.class.ts
