+++ meeting-room-management/src/app/statistic-room/model/booked-user.class.ts
