+++ meeting-room-management/src/app/statistic-room/model/room-type.class.ts
