+++ meeting-room-management/src/app/statistic-room/statistic-room.routers.ts
@@ -1,7 +1,11 @@
 import {Routes} from '@angular/router';
-import {ViewStatisticComponent} from './view-statistic/view-statistic.component';
+import {ViewStatisticComponent} from './component/view-statistic/view-statistic.component';
+import {AuthGuard} from '../office-common/helper/auth.guard';
 
 export const statisticRoomRoutes: Routes = [
   { path: 'statistic',
-    component: ViewStatisticComponent}
+    component: ViewStatisticComponent,
+    // canActivate: [AuthGuard],
+    // data: {roles: ['ROLE_ADMIN']}
+  }
 ];
